Changed ID type to BIGINT

+++ database/migrations/2025_06_17_162401_create_reports_table.php
@@ -14,18 +14,18 @@ return new class extends Migration
         Schema::disableForeignKeyConstraints();
 
         Schema::create('reports', function (Blueprint $table) {
-            $table->integer('id')->primary()->autoIncrement();
+            $table->bigInteger('id')->primary()->autoIncrement();
             $table->text('title');
             $table->enum('status', ["Reviewing","Investigating","Rejected","Resolving","Resolved"]);
             $table->integer('province_id')->comment('One-to-one');
             $table->foreign('province_id')->references('id')->on('provinces');
             $table->text('address');
             $table->longText('description');
-            $table->integer('citizen_id')->nullable()->comment('Associate report to a citizen. One report belong to one citizen only. One citizen can have many reports.');
+            $table->bigInteger('citizen_id')->nullable()->comment('Associate report to a citizen. One report belong to one citizen only. One citizen can have many reports.');
             $table->foreign('citizen_id')->references('id')->on('citizens');
             $table->timestamp('created_at');
             $table->timestamp('updated_at')->nullable();
-            $table->integer('updated_by')->nullable();
+            $table->bigInteger('updated_by')->nullable();
             $table->foreign('updated_by')->references('id')->on('officers');
             $table->text('remark')->nullable()->comment('Remark left by officers when they update status');
         });
